chore(data): add simulation scenarios
The scenarios belong to the coin-selection-simulation project
(https://github.com/achow101/coin-selection-simulation).

The files are two column csv with floating point numbers with up to
eight decimal places. Both express quantities in bitcoin.
The first column is signed, the sign of each number signals the nature
of the amount, if it is a deposit or a withdrawal.
The second column is integrated by the feerates (bitcoins per kilo
virtual byte) at which each deposit or withdrawal is done.