feat: add simulation runner
The only available algorithm to simulate for now is Branch and Bound
with a hard limit of 100.000 iterations, optimizing to get the lowest
overall fees: the one incurred at the moment of paying the fees to
register the transaction and the possible future ones associated to the
change outputs created.

Change is only created if it decreases the waste score of the selection.