refactor: flat conditionals inside simulation loop
By moving the deposits to the top of the loop and continuing after
adding them, the code can reduce one nesting level without compromising
its functionality.

The summary calculus should go before the liquidity check, to account
for the edge case where the final iteration is a withdrawal and there
isn't enough balance to pay it.