refactor(summary): unify conditionals for change stats
All change computed statistic values are related to the presence or not
of it. This is controlled by the value of the change, where a zero value
means no change and anything above zero is one or more change outputs.

This change unifies all that logic into a single conditional to improve
readability.