feat(selectors): define general behavior in trait
The goal of this simulator is to work as a wrapper for different coin
selection implementations in bitcoin which feeds each one of them with
generated scenarios. To work with heterogeneous implementations a common
interface to receive the products of the simulations should be
implemented.
TargetSelector, the trait implemented here, aims to fill that need.