fix(models): compose total_cost from total_fees and cost_to_empty
Total cost is the result of the total fees paid until the sample
snapshot and the cost to empty the utxo set at that very moment at the
long term fee rate.

Because of a typo, the formula was using the total cost (which was never
set so had a value of zero) instead of the total fees. This change fixes
this.